Last fixes beside conflict in dom.d.ts
